requirements.txt
flask
flask-cors
google-genai
pandas
torch
matplotlib
python-dotenv
